third_party/blink/web_tests/fast/writing-mode/block-formatting-context-expected.txt

The green LR block should avoid the blue float.
PASS: The floating block is in the correct position.
